Added ftsmooth, mkiv, patt & rfac Eclipse CDT projects with MinGW toolchain configuration.

+++ src/include/lattice.h
@@ -104,7 +104,7 @@ int lattice_set_script(lattice_t *lat, const char *script);
 void lattice_free(lattice_t *lat);
 void lattice_free_atom_list(lattice_t *lat);
 void lattice_read(lattice_t *lat, coord_t *a1, coord_t *a2, coord_t *a3, 
-                  coord_t *nor, coord_t *bas, char *bas_name, int *n_bas);
+                  coord_t *nor, coord_t *bas, char *bas_name, size_t *n_bas);
 int lattice_set_atom(lattice_t *lat, const atom_t *atom, size_t index);
 int lattice_set_atom_list(lattice_t *lat, const atom_t *atoms, size_t n_atoms);
 void lattice_printf(FILE *output, const lattice_t *lat);
@@ -131,7 +131,7 @@ coord_t *lattice_get_surface_normal(const lattice_t *lat, const coord_t *a1,
                                     const coord_t *a2, const coord_t *a3);
 double **lattice_normal_get_rotation_matrix(const coord_t *normal);
 void lattice_setup(lattice_t *lat, coord_t *a1, coord_t *a2, 
-         coord_t *a3, coord_t *nor, coord_t *bas, char *bas_name, int *n_bas);
+         coord_t *a3, coord_t *nor, coord_t *bas, char *bas_name, size_t *n_bas);
          
 miller_hkl_t *lattice_get_miller_hkl(const lattice_t *lat);
